perf: parallelize suspension handler for high-concurrency
- Process hooks first, then steps and waits in parallel to prevent race conditions
- Refactor runtime.ts into modular files: suspension-handler.ts, step-handler.ts, helpers.ts
- Add otel attributes for hooks created (workflow.hooks.created) and waits created (workflow.waits.created)
- Update suspension status from pending_steps to workflow_suspended
- Add retry logic to benchmark for unexpected content type errors
- Refactor stress test benchmarks to use shared config (10, 25 steps enabled; 100+ skipped)
